refactor(draft-loading-placeholder): rename scss files to use .module.scss
