+++ draft-packages/loading-placeholder/KaizenDraft/LoadingPlaceholder/_variables.scss
